refactor(api): modularize monolithic index.ts to separate concerns
[Why]
- The main entry point (`apps/api/src/index.ts`) had become a monolithic file (~160 lines) containing CORS settings, authentication, routing, and complex AI business logic.
- This structure increased technical debt, making the codebase harder to maintain, test, and scale for future API expansions.

[What]
- Extracted shared type definitions (`Bindings`, `Variables`) into `src/types.ts`.
- Modularized Supabase authentication logic into `src/middlewares/auth.ts`.
- Isolated Gemini API integration and quota management into a dedicated service layer (`src/services/ai.ts`).
- Created specific sub-routers for `/notes` and `/ai` using Hono's `app.route()` (`src/routes/notes.ts`, `src/routes/ai.ts`).
- Added survival tests (`src/index.test.ts`) to verify routing and access control.
- Ensured the existing API contracts (endpoints, request/response structures) remain strictly unchanged to prevent any impact on the frontend.
