+++ apps/api/package.json
@@ -12,6 +12,7 @@
 		"hono": "^4.11.6"
 	},
 	"devDependencies": {
+		"bun-types": "^1.3.12",
 		"wrangler": "^4.4.0"
 	}
 }
