+++ apps/api/tsconfig.json
@@ -7,6 +7,7 @@
 		"skipLibCheck": true,
 		"lib": ["ESNext"],
 		"jsx": "react-jsx",
-		"jsxImportSource": "hono/jsx"
+		"jsxImportSource": "hono/jsx",
+		"types": ["bun-types"]
 	}
 }
