+++ repomix-skeleton.config.json
@@ -1,0 +1,25 @@
+{
+	"include": [
+		".agent/**/*",
+		"package.json",
+		"turbo.json",
+		"**/package.json",
+		"**/tsconfig.json",
+		"**/README.md",
+		"**/*.d.ts",
+		"**/src/types/**/*.{ts,tsx}",
+		"**/src/constants/**/*.{ts,tsx}",
+		"**/src/index.{ts,tsx}",
+		"apps/app/src/app/**/page.tsx",
+		"apps/app/src/app/**/layout.tsx",
+		"apps/extension/src/entrypoints/**/index.tsx",
+		"supabase/schema.sql"
+	],
+	"output": {
+		"filePath": "repomix-skeleton-output.md",
+		"style": "markdown",
+		"removeComments": true,
+		"removeEmptyLines": true,
+		"topHeader": "PROJECT SKELETON: This file contains structure, rules (.agent/), and types. For implementation details of files not listed here, please ask for their content."
+	}
+}
